+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.8 - 2020-03-24
+* [enhancement] Support `Chat` target #65 [#65](https://github.com/treasure-data/embulk-input-zendesk/pull/65)
+* [enhancement] Apply Embulk Gradle  #64 [#64](https://github.com/treasure-data/embulk-input-zendesk/pull/64)
+
 ## 0.3.7 - 2019-08-29
 * [enhancement] Replace ConfigSource#getObjectNodes method in embulk core #63 [#63](https://github.com/treasure-data/embulk-input-zendesk/pull/63)